made changes according to monkey's review
