+++ listenbrainz/webserver/static/js/src/types.d.ts
@@ -580,9 +580,15 @@ type ColorReleasesResponse = {
     releases: Array<ColorReleaseItem>;
   };
 };
+
 type MissingMBData = {
   artist_name: string;
   listened_at: string;
   recording_name: string;
   release_name?: string;
+};
+
+type CustomIconFunction = {
+  icon: IconProp;
+  func: Function;
 };
